fastapi-llama-api/requirements.txt
requests
fastapi
uvicorn
youtube-transcript-api
groq
python-dotenv